+++ ru.kmorozov.gbd.core/src/main/java/ru/kmorozov/gbd/core/logic/connectors/asynchttp/AsyncHttpConnector.java
@@ -88,7 +88,7 @@ public class AsyncHttpConnector extends HttpConnector {
         }
 
         try {
-            return new AsyncHttpResponse(builder.execute(new AsyncHandler(proxy)).get(HttpConnector.CONNECT_TIMEOUT, TimeUnit.MILLISECONDS));
+            return new AsyncHttpResponse(builder.execute(new AsyncHandler(proxy)).get((long) HttpConnector.CONNECT_TIMEOUT, TimeUnit.MILLISECONDS));
         } catch (InterruptedException | ExecutionException | TimeoutException ex) {
             return null;
         } finally {
