+++ ru.kmorozov.gbd.core/src/main/java/ru/kmorozov/gbd/core/logic/connectors/webdriver/WebDriverHttpConnector.java
@@ -25,8 +25,8 @@ public class WebDriverHttpConnector extends HttpConnector {
         driver.get(url);
 
         final Wait<WebDriver> wait = new FluentWait(driver)
-                .withTimeout(CONNECT_TIMEOUT, TimeUnit.MILLISECONDS)
-                .pollingEvery(1, TimeUnit.SECONDS)
+                .withTimeout((long) CONNECT_TIMEOUT, TimeUnit.MILLISECONDS)
+                .pollingEvery(1L, TimeUnit.SECONDS)
                 .ignoring(NoSuchElementException.class);
 
         return new WebDriverResponse(driver.getPageSource());
