+++ ru.kmorozov.library.data.model/src/main/java/ru/kmorozov/library/utils/BookUtils.java
@@ -27,10 +27,10 @@ public class BookUtils {
 
     public static String humanReadableByteCount(final long bytes, final boolean si) {
         final int unit = si ? 1000 : 1024;
-        if (bytes < unit) return bytes + " B";
-        final int exp = (int) (Math.log(bytes) / Math.log(unit));
+        if (bytes < (long) unit) return bytes + " B";
+        final int exp = (int) (Math.log((double) bytes) / Math.log((double) unit));
         final String pre = (si ? "kMGTPE" : "KMGTPE").charAt(exp - 1) + (si ? "" : "i");
-        return String.format("%.1f %sB", bytes / Math.pow(unit, exp), pre);
+        return String.format("%.1f %sB", (double) bytes / Math.pow((double) unit, (double) exp), pre);
     }
 
     public static void mergeCategories(Book bookFrom, Book bookTo) {
